plant_disease_detection/deployment/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
firebase-admin
requests